+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/SendVideo.kt
@@ -92,7 +92,7 @@ data class SendVideoData internal constructor(
     @SerialName(replyMarkupField)
     @Optional
     override val replyMarkup: KeyboardMarkup? = null
-) : Data<RawMessage>,
+) : DataRequest<RawMessage>,
     SendMessageRequest<RawMessage>,
     ReplyingMarkupSendMessageRequest<RawMessage>,
     TextableSendMessageRequest<RawMessage>,