+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/MediaGroupList.kt
@@ -1,8 +1,10 @@
 package com.github.insanusmokrassar.TelegramBotAPI.utils
 
+import com.github.insanusmokrassar.TelegramBotAPI.types.MediaGroupIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.Chat
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.ForwardedMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.MediaGroupUpdate
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
 
 val List<BaseMessageUpdate>.forwarded: ForwardedMessage?
@@ -17,3 +19,6 @@ val List<BaseMessageUpdate>.replyTo: Message?
 
 val List<BaseMessageUpdate>.chat: Chat?
     get() = first().data.chat
+
+val List<BaseMessageUpdate>.mediaGroupId: MediaGroupIdentifier?
+    get() = (first().data as? MediaGroupMessage) ?.mediaGroupId